Version19. Formulario Recep.BrindarTurno trae todos los campos de la tabla pacientes, se ocultan mediante autocolumns:false el id_paciente y clave.

+++ TP_Cuatrimestral_Herrera_Scalesi/Dominio/Persona.cs
@@ -18,7 +18,7 @@ namespace Dominio
 
         public string Apellido { get; set; }
 
-        public int Dni { get; set; }
+        public short Dni { get; set; }
 
         public string Cuit { get; set; }
